+++ RegularExpression/OGRegularExpressionEnumeratorPrivate.m
@@ -120,12 +120,12 @@
 	_isLastMatchEmpty = yesOrNo;
 }
 
-- (void)_setStartLocation:(unsigned)location
+- (void)_setStartLocation:(NSUInteger)location
 {
 	_startLocation = location;
 }
 
-- (void)_setNumberOfMatches:(unsigned)aNumber
+- (void)_setNumberOfMatches:(NSUInteger)aNumber
 {
 	_numberOfMatches = aNumber;
 }
